Fix sum for 32-bit platform.

+++ sum/sum.go
@@ -2,6 +2,6 @@
 
 package sum
 
-func Sum(a, b int) int {
+func Sum(a, b int64) int64 {
 	return 0
 }
